✨ feat: add user identity and migration service - Implement CarRentalSystemUser and CarRentalSystemRole for identity management. - Create MigrationService for database migrations with OpenTelemetry support. - Update project references and configurations for new services.
