+++ CarRentalSystem.AppHost/AppHost.cs
@@ -5,11 +5,15 @@ var cache = builder.AddRedis("cache");
 var postgres = builder.AddPostgres("postgres")
     .AddDatabase("car-rental-system-db");
 
+var migrations = builder.AddProject<Projects.CarRentalSystem_MigrationService>("migrations")
+    .WithReference(postgres)
+    .WaitFor(postgres);
+
 var server = builder.AddProject<Projects.CarRentalSystem_Server>("server")
     .WithReference(postgres)
-    .WaitFor(postgres)
     .WithReference(cache)
-    .WaitFor(cache)
+    .WithReference(migrations)
+    .WaitFor(migrations)
     .WithHttpHealthCheck("/health")
     .WithExternalHttpEndpoints();
