+++ CarRentalSystem.Data/Models/CarRentalSystemRole.cs
@@ -1,0 +1,8 @@
+using Microsoft.AspNetCore.Identity;
+
+namespace CarRentalSystem.Data.Models;
+
+public class CarRentalSystemRole : IdentityRole
+{
+    
+}
