+++ CarRentalSystem.Data/Models/CarRentalSystemUser.cs
@@ -1,0 +1,7 @@
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace CarRentalSystem.Data.Models;
+
+public class CarRentalSystemUser : IdentityUser
+{
+}
