+++ CarRentalSystem.MigrationService/CarRentalSystem.MigrationService.csproj
@@ -1,0 +1,19 @@
+<Project Sdk="Microsoft.NET.Sdk.Worker">
+
+    <PropertyGroup>
+        <TargetFramework>net10.0</TargetFramework>
+        <Nullable>enable</Nullable>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <UserSecretsId>dotnet-CarRentalSystem.MigrationService-caf70d18-ddc0-4bba-9cc5-9636b6aa55ac</UserSecretsId>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Aspire.Npgsql.EntityFrameworkCore.PostgreSQL" Version="13.1.2" />
+        <PackageReference Include="Microsoft.Extensions.Hosting" Version="10.0.3"/>
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\CarRentalSystem.Data\CarRentalSystem.Data.csproj" />
+      <ProjectReference Include="..\CarRentalSystem.ServiceDefaults\CarRentalSystem.ServiceDefaults.csproj" />
+    </ItemGroup>
+</Project>
