+++ CarRentalSystem.MigrationService/MigrationWorker.cs
@@ -1,0 +1,39 @@
+using System.Diagnostics;
+using CarRentalSystem.Data.Contexts;
+using Microsoft.EntityFrameworkCore;
+
+namespace CarRentalSystem.MigrationService;
+
+public class MigrationWorker(IServiceProvider serviceProvider, IHostApplicationLifetime hostApplicationLifetime) : BackgroundService
+{
+    public const string ActivitySourceName = "Migrations";
+    private static readonly ActivitySource s_activitySource = new(ActivitySourceName);
+    
+    protected override async Task ExecuteAsync(CancellationToken cancellationToken)
+    {
+        using var activity = s_activitySource.StartActivity(
+            "Migrating database", ActivityKind.Client);
+
+        try
+        {
+            using var scope = serviceProvider.CreateScope();
+            var dbContext = scope.ServiceProvider.GetRequiredService<CarRentalSystemDbContext>();
+
+            await RunMigrationAsync(dbContext, cancellationToken);
+        }
+        catch (Exception ex)
+        {
+            activity?.AddException(ex);
+            throw;
+        }
+    }
+
+    private static async Task RunMigrationAsync(CarRentalSystemDbContext dbContext, CancellationToken cancellationToken)
+    {
+        var strategy = dbContext.Database.CreateExecutionStrategy();
+        await strategy.ExecuteAsync(async () =>
+        {
+            await dbContext.Database.MigrateAsync(cancellationToken);
+        });
+    }
+}
