+++ CarRentalSystem.sln
@@ -10,6 +10,10 @@ Project("{54A90642-561A-4BB1-A94E-469ADEE60C69}") = "frontend", "frontend\fronte
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CarRentalSystem.Data", "CarRentalSystem.Data\CarRentalSystem.Data.csproj", "{8F13D26C-BC7A-472C-8EEC-FFF81957DF91}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CarRentalSystem.MigrationService", "CarRentalSystem.MigrationService\CarRentalSystem.MigrationService.csproj", "{B67E6824-2551-46D3-B878-C98A69E6B582}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CarRentalSystem.ServiceDefaults", "CarRentalSystem.ServiceDefaults\CarRentalSystem.ServiceDefaults.csproj", "{45C1FB3D-768A-4C87-BE1F-D63A60539CF3}"
+EndProject
 
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -34,6 +38,14 @@ Global
 		{8F13D26C-BC7A-472C-8EEC-FFF81957DF91}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8F13D26C-BC7A-472C-8EEC-FFF81957DF91}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8F13D26C-BC7A-472C-8EEC-FFF81957DF91}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B67E6824-2551-46D3-B878-C98A69E6B582}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B67E6824-2551-46D3-B878-C98A69E6B582}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B67E6824-2551-46D3-B878-C98A69E6B582}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B67E6824-2551-46D3-B878-C98A69E6B582}.Release|Any CPU.Build.0 = Release|Any CPU
+		{45C1FB3D-768A-4C87-BE1F-D63A60539CF3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{45C1FB3D-768A-4C87-BE1F-D63A60539CF3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{45C1FB3D-768A-4C87-BE1F-D63A60539CF3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{45C1FB3D-768A-4C87-BE1F-D63A60539CF3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
